+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Previous classification is not required if changes are simple or all belong to t
   - Combines strict-format PDF text extraction with image analysis and interpretation using the vision model.
   - Detects embedded images in PDF files, processes them using vision, and appends their descriptions as additional text blocks.
 - Introduced a new virtual extension point `ProcessPageExtensions(Page page)` in `StrictFormatCleanPdfDocumentConnector`, allowing derived classes to append custom content blocks to the extracted page content.
+- Added a new connector `SkVisionWordDocumentConnector` to extract text, tables, hyperlinks and images from Word documents (.docx) using Semantic Kernel vision capabilities in image extractions. This capacity to extract the description for each image is realized using the existing `SkVisionImageExtractor`. This functionality is compatible with the previous `WordDocumentConector` implementation, so it can be used interchangeably using the `WordDocumentConnector:UseImageExtractor` setting.
 
 - Updated dependencies:
   - Aspire.Hosting 8.2.1 → 8.2.2